Search fixes on save, search, refine.
git-svn-id: https://svn.java.net/svn/docsources~svn/trunk@946 3b25252b-017b-4ae5-8f61-fa87a855a72e

+++ DocSources/src/main/java/org/medici/docsources/domain/SearchFilter.java
@@ -77,7 +77,7 @@ public class SearchFilter implements Serializable {
 	private AdvancedSearch filterData;
 
 	@Column (name="\"totalResult\"", nullable=false)
-	private Integer totalResult;
+	private Long totalResult;
 	
 	@Column (name="\"dateCreated\"", nullable=false)
 	@Temporal(TemporalType.TIMESTAMP)
@@ -174,14 +174,14 @@ public class SearchFilter implements Serializable {
 	/**
 	 * @param totalResult the totalResult to set
 	 */
-	public void setTotalResult(Integer totalResult) {
+	public void setTotalResult(Long totalResult) {
 		this.totalResult = totalResult;
 	}
 
 	/**
 	 * @return the totalResult
 	 */
-	public Integer getTotalResult() {
+	public Long getTotalResult() {
 		return totalResult;
 	}
